rename views
