+++ src/main.js
@@ -2,7 +2,7 @@ import TripPresenter from './presenter/trip-presenter';
 import PointModel from './model/point-model';
 import {generateFilter} from './mock/filter';
 import {render} from './framework/render';
-import FiltersView from './view/filters';
+import FiltersView from './view/filters-view';
 
 
 const siteHeaderElement = document.querySelector('.trip-main');
