+++ src/view/empty-points-list-view.js
